requirements.txt
fortigate-api
keyring
#win32crypt
pycryptodome
#pypiwin32
pyyaml